Write all of the main interaction with the user
Inside the VendingMachine Class I have all the user interfacing, and withods to complete the base application. All that is left to impletemt is logging capability and the hidden menu for sales. Also to finish off the project I want to add it to a webpage and have the java code called to from the page